Moved form to components folder
